Add a manifest for ARM64 kube-flannel daemonset.
Since the current logic relies on the node being master['0']
we can use include to wait for a node that matches the arm64
architecture. Once/if we find one, we will include the task that
creates the arm64 manifest for kube-flannel. This could possibly
be improved with some jinja2 logic. But again, since we're running
only on the master. We would need to write logic that would check
the inventory for any ARM64 devices and then render the appropriate
manifests.